Use absolute import in tsconfig.json. Run dotenv in config/. Rename util/ to utils/
