+++ tsconfig.json
@@ -7,10 +7,17 @@
     "noImplicitAny": false, // Might be overly stringent if this is set to true. Exercise judgement here
     "moduleResolution": "node",
     "sourceMap": true,
-    "outDir": "dist",
-    "baseUrl": ".",
+    "outDir": "./dist",
+    "baseUrl": "./src",
     "paths": {
-      "*": ["node_modules/*", "src/types/*"]
+      "config": ["config/*"],
+      "controllers": ["controllers/*"],
+      "db": ["db/*"],
+      "middlewares": ["middlewares/*"],
+      "models": ["models/*"],
+      "routes": ["routes/*"],
+      "utils": ["utils/*"],
+      "types": ["types/*"]
     }
   },
   "include": ["src/**/*"]
